# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project Ledger API",
    "version": "1.0.0",
    "description": "JSON API for projects, RAID records, and changelog entries. No authentication is required; all visitors can read, create, edit, and delete records. PATCH updates supplied fields only. Project deletion cascades to its children."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current application host"
    }
  ],
  "security": [],
...
  "paths": {
    "/api/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "listProjects",
        "summary": "List projects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Records ordered by descending ID; empty collections return an empty array.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "projects"
                  ],
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create Project",
        "operationId": "postProject",
        "parameters": [],
        "responses": {
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "201": {
            "description": "Project with all associated RAID records and changelog entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "project"
                  ],
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/ProjectDetail"
                    }
                  }
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL of the created record.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "getProject",
        "summary": "Get project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "description": "ID of the owning project.",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Project with all associated RAID records and changelog entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "project"
                  ],
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/ProjectDetail"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "operationId": "patchProject",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "description": "ID of the owning project.",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "200": {
            "description": "Project with all associated RAID records and changelog entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "project"
                  ],
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/ProjectDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete Project",
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "description": "ID of the owning project.",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "204": {
            "description": "Deleted. Empty response body. Project deletion also deletes its RAID and changelog records."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "example": "Website launch"
          },
          "description": {
            "type": "string",
            "example": "Launch the new website."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "UTC creation timestamp.",
            "example": "2026-09-25T12:00:00Z"
          }
        }
      },
      "RaidItem": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "kind",
          "title",
          "description",
          "owner",
          "status",
          "priority",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "project_id": {
            "type": "integer",
            "example": 1
          },
          "kind": {
            "type": "string",
            "enum": [
              "Risk",
              "Assumption",
              "Issue",
              "Dependency"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "Vendor delivery delay"
          },
          "description": {
            "type": "string"
          },
          "owner": {
            "type": "string",
            "maxLength": 200
          },
          "status": {
            "type": "string",
            "enum": [
              "Open",
              "In progress",
              "Closed"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "Low",
              "Medium",
              "High"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "UTC creation timestamp.",
            "example": "2026-09-25T12:00:00Z"
          }
        }
      },
      "ChangelogEntry": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "body",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "project_id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "Kickoff completed"
          },
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "UTC creation timestamp.",
            "example": "2026-09-25T12:00:00Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "integer",
                "example": 404
              },
              "message": {
                "type": "string",
                "example": "The requested URL was not found on the server."
              }
            }
          }
        }
      },
      "ProjectDetail": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "raid_items",
          "changelog_entries"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "example": "Website launch"
          },
          "description": {
            "type": "string",
            "example": "Launch the new website."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "UTC creation timestamp.",
            "example": "2026-09-25T12:00:00Z"
          },
          "raid_items": {
            "type": "array",
            "description": "All associated records, newest ID first. Empty when no records exist.",
            "items": {
              "$ref": "#/components/schemas/RaidItem"
            }
          },
          "changelog_entries": {
            "type": "array",
            "description": "All associated records, newest ID first. Empty when no records exist.",
            "items": {
              "$ref": "#/components/schemas/ChangelogEntry"
            }
          }
        }
      },
      "ProjectCreate": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "example": "Website launch",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "example": "Launch the new website.",
            "maxLength": 10000
          }
        },
        "required": [
          "name"
        ]
      },
      "ProjectUpdate": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "example": "Website launch",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "example": "Launch the new website.",
            "maxLength": 10000
          }
        },
        "required": []
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "Project or record not found, or record belongs to a different project.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "HTTP method is not supported for this URL.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error400": {
        "description": "Invalid JSON or field values.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error415": {
        "description": "Content-Type must be application/json.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}